+++ test/networks.js
@@ -54,7 +54,7 @@ describe("Different networks:", function() {
     this.timeout(10000);
 
     // Compile first
-    var result = solc.compile(fs.readFileSync("./test/Example.sol", {encoding: "utf8"}), 1);
+    var result = solc.compile(fs.readFileSync("./test/sources/Example.sol", {encoding: "utf8"}), 1);
 
     var compiled = Schema.normalize(
       result.contracts["Example"] || result.contracts[":Example"]
